developed unit tests for create product

+++ src/main/java/com/novik/productsapi/dto/request/ProductDTO.java
@@ -1,7 +1,10 @@
 package com.novik.productsapi.dto.request;
 
 import com.novik.productsapi.enums.Inventory;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
@@ -9,8 +12,8 @@ import javax.validation.Valid;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
-import java.math.BigDecimal;
 import java.util.List;
+
 
 @Data
 @Builder
@@ -18,7 +21,7 @@ import java.util.List;
 @NoArgsConstructor
 public class ProductDTO {
 
-    private Long id;
+   private Long id;
 
     @NotEmpty
     @Size(min = 2, max = 25)
@@ -36,10 +39,10 @@ public class ProductDTO {
     private int amount;
 
     @NotNull
-    private BigDecimal weight;
+    private double weight;
 
     @NotNull
-    private BigDecimal unitValue;
+    private double unitValue;
 
     @NotEmpty
     private String maturityDate;
@@ -53,3 +56,5 @@ public class ProductDTO {
 
 
 }
+
+
